split image & video processing
